Fix duplicate union check
The purpose of the check was to prevent redundant work for the rules that
navigate through `Union`s, as they already handle nested `Union`s. But this
was not happening.

If we had the following code:
```python
typing.Union[Literal[5], int, typing.Union[Literal["foo"], str]]
```
The rules were receiving the expressions in the following order:
 - `typing.Union[Literal[5], int, typing.Union[Literal["foo"], str]]`
 - `Literal[5]`
 - `typing.Union[Literal["foo"], str]]`

Which is wrong, they are only supposed to receive the outermost `Union`
but with the previous check they were receiving even a `Literal`.

The fix is, now we check if the current `expr` has no parent (which is true
if we are the outermost `Union`) and the `expr` is an `Union`.