Fix formatting of `lambda` star arguments
Previously, the ruff formatter was removing the star argument of `lambda`
expressions when formatting.

Given the following code snippet
```python
lambda *a: ()
lambda **b: ()
```
it would be formated to
```python
lambda: ()
lambda: ()
```

We fix this by checking for the presence of `args`, `vararg` or `kwarg` in the
`lambda` expression, before we were only checking  for the presence of `args`.

Fixes #5894