move comments from expressions in f-strings out
move comments from expressions inside f-strings out to the containing
f-string.

py<3.12 doesn't support comments inside formatted values. once
we support 3.12+ f-strings we may need to revisit this. i suspect we'll
end up with chunks of different behaviour for pre vs post 3.12 f-strings
and this will be part of that.